pyproject.toml
[tool.poetry]
name = "multi-agent-deep-researcher"
version = "0.1.0"
description = "MCP-powered multi-agent deep researcher using CrewAI, LinkUp, and Deepseek R1."
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
crewai = "^0.141.0"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
pydantic = "^2.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 